rename current employment section
